apps/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pydantic-settings==2.6.1
jinja2==3.1.6
python-multipart==0.0.9
fastapi-mail==1.6.2
